refactor: single file components and hooks
