+++ src/components/RepositoryPages/styled.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const PageSelectors = styled.div`
+  margin: 0 auto;
+  text-align: center;
+`;
