+++ src/utils/useDebounce.js
@@ -1,0 +1,16 @@
+import { useState, useEffect } from "react";
+
+function useDebounce(query, delay) {
+  const [state, setState] = useState(query);
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      return setState(query);
+    }, 250);
+    return () => clearTimeout(timer);
+  }, [query, delay]);
+
+  return state;
+}
+
+export default useDebounce;
